Move social security number input to ui-core module
COAND-366
