+++ ui-core/src/main/res/values-cs-rCZ/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Zrušit</string>
 
     <string name="checkout_address_form_field_not_valid">Neplatný vstup</string>
+    <string name="checkout_social_security_number_not_valid">Neplatný vstup</string>
 
     <string name="checkout_address_form_billing_address_title">Fakturační adresa</string>
     <string name="checkout_address_form_street_hint">Ulice</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">PSČ (nepovinné)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Provincie nebo teritorium</string>
     <string name="checkout_address_form_city_town_hint_optional">Město (nepovinné)</string>
-
 </resources>
