+++ ui-core/src/main/res/values-da-rDK/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Annuller</string>
 
     <string name="checkout_address_form_field_not_valid">Ugyldig indtastning</string>
+    <string name="checkout_social_security_number_not_valid">Ugyldig indtastning</string>
 
     <string name="checkout_address_form_billing_address_title">Faktureringsadresse</string>
     <string name="checkout_address_form_street_hint">Gade</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Postnummer (valgfrit)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Provins eller territorium (valgfrit)</string>
     <string name="checkout_address_form_city_town_hint_optional">By (valgfrit)</string>
-
 </resources>
