+++ ui-core/src/main/res/values-de-rDE/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Abbrechen</string>
 
     <string name="checkout_address_form_field_not_valid">Ungültige Eingabe</string>
+    <string name="checkout_social_security_number_not_valid">Ungültige Eingabe</string>
 
     <string name="checkout_address_form_billing_address_title">Rechnungsadresse</string>
     <string name="checkout_address_form_street_hint">Straße</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">PLZ (optional)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Provinz oder Territorium (optional)</string>
     <string name="checkout_address_form_city_town_hint_optional">Ort (optional)</string>
-
 </resources>
