+++ ui-core/src/main/res/values-el-rGR/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Άκυρο</string>
 
     <string name="checkout_address_form_field_not_valid">Μη έγκυρη είσοδος</string>
+    <string name="checkout_social_security_number_not_valid">Μη έγκυρη είσοδος</string>
 
     <string name="checkout_address_form_billing_address_title">Διεύθυνση τιμολόγησης</string>
     <string name="checkout_address_form_street_hint">Οδός</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Ταχυδρομικός κωδικός (προαιρετικό)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Επαρχία ή περιοχή (προαιρετικό)</string>
     <string name="checkout_address_form_city_town_hint_optional">Πόλη / Δήμος (προαιρετικό)</string>
-
 </resources>
