+++ ui-core/src/main/res/values-fr-rFR/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Annuler</string>
 
     <string name="checkout_address_form_field_not_valid">Saisie incorrecte</string>
+    <string name="checkout_social_security_number_not_valid">Saisie incorrecte</string>
 
     <string name="checkout_address_form_billing_address_title">Adresse de facturation</string>
     <string name="checkout_address_form_street_hint">Rue</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Code postal (facultatif)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Province ou territoire (facultatif)</string>
     <string name="checkout_address_form_city_town_hint_optional">Ville (facultatif)</string>
-
 </resources>
