+++ ui-core/src/main/res/values-hr-rHR/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Otkaži</string>
 
     <string name="checkout_address_form_field_not_valid">Pogrešan unos</string>
+    <string name="checkout_social_security_number_not_valid">Pogrešan unos</string>
 
     <string name="checkout_address_form_billing_address_title">Adresa za račun</string>
     <string name="checkout_address_form_street_hint">Ulica</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Poštanski broj (neobavezno)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Pokrajina ili teritorij (neobavezno)</string>
     <string name="checkout_address_form_city_town_hint_optional">Grad (neobavezno)</string>
-
 </resources>
