+++ ui-core/src/main/res/values-hu-rHU/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Mégse</string>
 
     <string name="checkout_address_form_field_not_valid">Érvénytelen bevitel</string>
+    <string name="checkout_social_security_number_not_valid">Érvénytelen bevitel</string>
 
     <string name="checkout_address_form_billing_address_title">Számlázási cím</string>
     <string name="checkout_address_form_street_hint">Utca</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Irányítószám (nem kötelező)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Tartomány vagy terület (nem kötelező)</string>
     <string name="checkout_address_form_city_town_hint_optional">Város (nem kötelező)</string>
-
 </resources>
