+++ ui-core/src/main/res/values-it-rIT/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Cancella</string>
 
     <string name="checkout_address_form_field_not_valid">Dati inseriti non validi</string>
+    <string name="checkout_social_security_number_not_valid">Dati inseriti non validi</string>
 
     <string name="checkout_address_form_billing_address_title">Indirizzo di fatturazione</string>
     <string name="checkout_address_form_street_hint">Via</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">CAP (facoltativo)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Provincia o territorio (facoltativo)</string>
     <string name="checkout_address_form_city_town_hint_optional">Città (facoltativo)</string>
-
 </resources>
