+++ ui-core/src/main/res/values-ko-rKR/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">취소</string>
 
     <string name="checkout_address_form_field_not_valid">유효하지 않은 입력</string>
+    <string name="checkout_social_security_number_not_valid">유효하지 않은 입력</string>
 
     <string name="checkout_address_form_billing_address_title">청구지 주소</string>
     <string name="checkout_address_form_street_hint">도로명</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">우편 번호(선택 사항)</string>
     <string name="checkout_address_form_province_territory_hint_optional">도(선택 사항)</string>
     <string name="checkout_address_form_city_town_hint_optional">시/구(선택 사항)</string>
-
 </resources>
