+++ ui-core/src/main/res/values-nl-rNL/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Annuleren</string>
 
     <string name="checkout_address_form_field_not_valid">Ongeldige invoer</string>
+    <string name="checkout_social_security_number_not_valid">Ongeldige invoer</string>
 
     <string name="checkout_address_form_billing_address_title">Factuuradres</string>
     <string name="checkout_address_form_street_hint">Straatnaam</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Postcode (optioneel)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Provincie of territorium (optioneel)</string>
     <string name="checkout_address_form_city_town_hint_optional">Stad (optioneel)</string>
-
 </resources>
