+++ ui-core/src/main/res/values-pl-rPL/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Anuluj</string>
 
     <string name="checkout_address_form_field_not_valid">Nieprawidłowe dane wejściowe</string>
+    <string name="checkout_social_security_number_not_valid">Nieprawidłowe dane wejściowe</string>
 
     <string name="checkout_address_form_billing_address_title">Adres rozliczeniowy</string>
     <string name="checkout_address_form_street_hint">Ulica</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Kod pocztowy (opcjonalnie)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Region lub terytorium (opcjonalnie)</string>
     <string name="checkout_address_form_city_town_hint_optional">Miejscowość (opcjonalnie)</string>
-
 </resources>
