+++ ui-core/src/main/res/values-pt-rBR/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Cancelar</string>
 
     <string name="checkout_address_form_field_not_valid">Entrada inválida</string>
+    <string name="checkout_social_security_number_not_valid">Entrada inválida</string>
 
     <string name="checkout_address_form_billing_address_title">Endereço de cobrança</string>
     <string name="checkout_address_form_street_hint">Rua</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Código postal (opcional)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Província ou território (opcional)</string>
     <string name="checkout_address_form_city_town_hint_optional">Cidade (opcional)</string>
-
 </resources>
