+++ ui-core/src/main/res/values-ro-rRO/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Anulare</string>
 
     <string name="checkout_address_form_field_not_valid">Informațiile sunt incorecte</string>
+    <string name="checkout_social_security_number_not_valid">Informațiile sunt incorecte</string>
 
     <string name="checkout_address_form_billing_address_title">Adresa de facturare</string>
     <string name="checkout_address_form_street_hint">Strada</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Cod poștal (opțional)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Provincie sau teritoriu (opțional)</string>
     <string name="checkout_address_form_city_town_hint_optional">Oraș/localitate (opțional)</string>
-
 </resources>
