+++ ui-core/src/main/res/values-ru-rRU/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Отменить</string>
 
     <string name="checkout_address_form_field_not_valid">Неверные данные</string>
+    <string name="checkout_social_security_number_not_valid">Неверные данные</string>
 
     <string name="checkout_address_form_billing_address_title">Платежный адрес</string>
     <string name="checkout_address_form_street_hint">Улица</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Почтовый индекс (необязательно)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Область или территория (необязательно)</string>
     <string name="checkout_address_form_city_town_hint_optional">Город (необязательно)</string>
-
 </resources>
