+++ ui-core/src/main/res/values-sl-rSI/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Prekliči</string>
 
     <string name="checkout_address_form_field_not_valid">Neveljaven vnos</string>
+    <string name="checkout_social_security_number_not_valid">Neveljaven vnos</string>
 
     <string name="checkout_address_form_billing_address_title">Naslov za račun</string>
     <string name="checkout_address_form_street_hint">Ulica</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Poštna številka (neobvezno)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Območje ali ozemlje (neobvezno)</string>
     <string name="checkout_address_form_city_town_hint_optional">Mesto (neobvezno)</string>
-
 </resources>
