+++ ui-core/src/main/res/values-sv-rSE/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Avbryt</string>
 
     <string name="checkout_address_form_field_not_valid">Ogiltig inmatning</string>
+    <string name="checkout_social_security_number_not_valid">Ogiltig inmatning</string>
 
     <string name="checkout_address_form_billing_address_title">Faktureringsadress</string>
     <string name="checkout_address_form_street_hint">Gatuadress</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">Postnummer (valfritt)</string>
     <string name="checkout_address_form_province_territory_hint_optional">Provins eller territorium (valfritt)</string>
     <string name="checkout_address_form_city_town_hint_optional">Ort (valfritt)</string>
-
 </resources>
