+++ ui-core/src/main/res/values-zh-rCN/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">取消</string>
 
     <string name="checkout_address_form_field_not_valid">无效的输入</string>
+    <string name="checkout_social_security_number_not_valid">无效的输入</string>
 
     <string name="checkout_address_form_billing_address_title">账单地址</string>
     <string name="checkout_address_form_street_hint">街道</string>
@@ -41,5 +42,4 @@
     <string name="checkout_address_form_zip_code_hint_optional">邮政编码（可选）</string>
     <string name="checkout_address_form_province_territory_hint_optional">省或地区（可选）</string>
     <string name="checkout_address_form_city_town_hint_optional">市 / 镇（可选）</string>
-
 </resources>
