+++ ui-core/src/main/res/values/strings.xml
@@ -16,6 +16,7 @@
     <string name="paymentInProgressView_cancelButton">Cancel</string>
 
     <string name="checkout_address_form_field_not_valid">Invalid Input</string>
+    <string name="checkout_social_security_number_not_valid">Invalid Input</string>
 
     <string name="checkout_address_form_billing_address_title">Billing address</string>
     <string name="checkout_address_form_street_hint">Street</string>
